feat: Firebase 연동 관련 코드 분리
